Create lines module with a refund and sale model to separate logic and refactor
